Relocate Flet modules and update imports
