+++ .coveragerc
@@ -1,4 +1,4 @@
 [run]
 omit =
-    src/flet_app.py
+    src/genecoder/flet_app.py
     src/cli.py
